Refactor scrollbar util to be part of mask feature
