+++ src/features/mask/scrollbar.ts
@@ -1,4 +1,4 @@
-import { clearCache, fromCache } from './cache';
+import { clearCache, fromCache } from '../../utils/cache';
 
 
 export type ScrollbarDimensions = {
@@ -12,6 +12,7 @@ export class Scrollbar {
 
 	/**
 	 * Creates an instance.
+	 * @internal
 	 */
 	constructor() {
 		window.addEventListener('resize', () => {
